// src/types/chances.ts
class Chances {
  sampleSize: number;

  landW: number;

  landU: number;

  landB: number;

  landR: number;

  landG: number;

  chanceCre: number;

  chanceIns: number;

  chanceSor: number;

  chancePla: number;

  chanceArt: number;

  chanceEnc: number;

  chanceBat: number;

  chanceLan: number;

  deckSize: number;

  cardsLeft: number;

  constructor() {
    this.sampleSize = 0;
    this.landW = 0;
    this.landU = 0;
    this.landB = 0;
    this.landR = 0;
    this.landG = 0;
    this.chanceCre = 0;
    this.chanceIns = 0;
    this.chanceSor = 0;
    this.chancePla = 0;
    this.chanceArt = 0;
    this.chanceEnc = 0;
    this.chanceBat = 0;
    this.chanceLan = 0;
    this.deckSize = 0;
    this.cardsLeft = 0;
  }
}

export default Chances;